Change representation of interfaces.
This makes them more like the gc compiler.  An interface is
now a struct of two pointers.  For an empty interface the
first pointer is the type descriptor.  For a non-empty
interface the first pointer points to the method table, and
the first entry in the method table is the type descriptor.

R=iant
http://codereview.appspot.com/1715055