compiler: don't export embedded builtins.
The panic in test/fixedbugs/bug461.go was caused by the fact that
reflect expects unexported fields in a struct to have a valid
package path. If a struct field is an embedded built-in type, it is
now given the package name of the currently compiling package, so it
remains unexported for purposes of reflect.

Fixed Issue 25.

R=gofrontend-dev, iant
CC=gofrontend-dev
https://codereview.appspot.com/12815051

Committer: Ian Lance Taylor <iant@golang.org>