compiler: Don't convert function type for an interface method.
PR go/60870

For an interface method the function type is the type without
the receiver, which is wrong since we are passing a receiver.
The interface method will always have the correct type in this
case, so no type conversion is necessary.

Also don't do the type conversion when calling a named
function, since in that case the type is also always correct.

The type can be wrong, and the conversion required, when the
function type refers to itself recursively.

R=iant
CC=gofrontend-dev
https://codereview.appspot.com/89010043