compiler: Do not count package uses from ambiguous lookups.
When using dot imports, it is possible to
have an imported symbol name that matches
the name of a struct field in a composite
literal.  Do not consider the imported
package to be used in this situation.
Fixes issue 6427.

LGTM=iant
R=iant
CC=gofrontend-dev
https://codereview.appspot.com/159580043

Committer: Ian Lance Taylor <iant@golang.org>